Harden harness default test coverage (#498)
Use observable GPUI window state and stable panic-context suffixes to
exercise the runtime contracts directly. Separate dependency-facing test
support from path-included panic helpers so plain and all-feature builds
share sound crate boundaries.

Propagate lint-runner I/O failures and update the documented validation
commands and Mermaid fixture labels.
