+++ crates/rstest-bdd-harness/src/test_utils.rs
@@ -29,24 +29,3 @@ pub(crate) fn panic_payload_matches(payload: &(dyn Any + Send), expected: &str) 
             .downcast_ref::<String>()
             .is_some_and(|message| message == expected)
 }
-
-/// A harness whose `run` always returns
-/// `HarnessError::RuntimeBuildFailed` with a synthetic IO error, for use in
-/// test suites that verify error-path behaviour.
-#[cfg(feature = "testing")]
-#[derive(Default)]
-pub struct FailingHarness;
-
-#[cfg(feature = "testing")]
-impl crate::HarnessAdapter for FailingHarness {
-    type Context = ();
-
-    fn run<T>(
-        &self,
-        _request: crate::ScenarioRunRequest<'_, Self::Context, T>,
-    ) -> crate::HarnessResult<T> {
-        Err(crate::HarnessError::RuntimeBuildFailed(
-            std::io::Error::other("synthetic harness initialization failure"),
-        ))
-    }
-}
